--- a/flink-core/src/main/java/org/apache/flink/core/memory/DataInputDeserializer.java
+++ b/flink-core/src/main/java/org/apache/flink/core/memory/DataInputDeserializer.java
@@ -64,7 +64,8 @@ public class DataInputDeserializer implements DataInputView, java.io.Serializabl
 			this.buffer = buffer.array();
 			this.position = buffer.arrayOffset() + buffer.position();
 			this.end = this.position + buffer.remaining();
-		} else if (buffer.isDirect()) {
+		} else if (buffer.isDirect() || buffer.isReadOnly()) {
+			// TODO: FLINK-8585 handle readonly and other non array based buffers more efficiently without data copy
 			this.buffer = new byte[buffer.remaining()];
 			this.position = 0;
 			this.end = this.buffer.length;