--- a/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/api/serialization/RecordDeserializer.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/api/serialization/RecordDeserializer.java
@@ -19,7 +19,6 @@
 package org.apache.flink.runtime.io.network.api.serialization;
 
 import org.apache.flink.core.io.IOReadableWritable;
-import org.apache.flink.core.memory.MemorySegment;
 import org.apache.flink.runtime.io.network.buffer.Buffer;
 
 import java.io.IOException;
@@ -56,8 +55,6 @@ public interface RecordDeserializer<T extends IOReadableWritable> {
 	}
 
 	DeserializationResult getNextRecord(T target) throws IOException;
-
-	void setNextMemorySegment(MemorySegment segment, int numBytes) throws IOException;
 
 	void setNextBuffer(Buffer buffer) throws IOException;