--- a/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/api/serialization/RecordSerializer.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/api/serialization/RecordSerializer.java
@@ -83,7 +83,7 @@ public interface RecordSerializer<T extends IOReadableWritable> {
 	 * @return how much information was written to the target buffer and
 	 *         whether this buffer is full
 	 */
-	SerializationResult setNextBufferBuilder(BufferBuilder bufferBuilder) throws IOException;
+	SerializationResult continueWritingWithNextBufferBuilder(BufferBuilder bufferBuilder) throws IOException;
 
 	/**
 	 * Clear and release internal state.