--- a/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/api/writer/RecordWriter.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/api/writer/RecordWriter.java
@@ -132,7 +132,7 @@ public class RecordWriter<T extends IOReadableWritable> {
 			}
 			BufferBuilder bufferBuilder = requestNewBufferBuilder(targetChannel);
 
-			result = serializer.setNextBufferBuilder(bufferBuilder);
+			result = serializer.continueWritingWithNextBufferBuilder(bufferBuilder);
 		}
 		checkState(!serializer.hasSerializedData(), "All data should be written at once");
 	}