… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/api/serialization/AdaptiveSpanningRecordDeserializer.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/buffer/Buffer.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/buffer/BufferBuilder.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/buffer/BufferConsumer.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/buffer/LocalBufferPool.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/buffer/NetworkBuffer.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/buffer/ReadOnlySlicedNetworkBuffer.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/netty/CreditBasedSequenceNumberingViewReader.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/netty/PartitionRequestQueue.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/netty/SequenceNumberingViewReader.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/PipelinedSubpartition.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/PipelinedSubpartitionView.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/ResultPartition.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/ResultSubpartition.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/ResultSubpartitionView.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/SpillableSubpartition.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/SpillableSubpartitionView.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/SpilledSubpartitionView.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/consumer/InputChannel.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/consumer/InputGate.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/consumer/LocalInputChannel.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/consumer/RemoteInputChannel.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/consumer/SingleInputGate.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/consumer/UnionInputGate.java
… flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/consumer/UnknownInputChannel.java
… flink-runtime/src/main/java/org/apache/flink/runtime/iterative/task/IterationHeadTask.java
… flink-runtime/src/main/java/org/apache/flink/runtime/operators/shipping/OutputCollector.java
… flink-runtime/src/test/java/org/apache/flink/runtime/io/network/api/serialization/SpanningRecordSerializationTest.java
… flink-runtime/src/test/java/org/apache/flink/runtime/io/network/api/serialization/SpanningRecordSerializerTest.java
… flink-runtime/src/test/java/org/apache/flink/runtime/io/network/api/writer/AbstractCollectingResultPartitionWriter.java