This big PR depends on #4552 and #5314. Main purpose of this change is to increase network throughput/performance in low latency cases. On its own, #4552 and #5314 are causing huge performance degradation for ~1ms flushing intervals (on top of already very poor Flink's performance in such case). This PR is fixing making throughput in ~1ms more or less similar to ~100ms flushing interval.

Quick (noisy) benchmark results:

master branch:
```
Benchmark    (number of output channels, flush interval)   Mode  Cnt      Score       Error   Units
networkThroughput                 1,100ms  thrpt    5  53776.816 ± 8566.861  ops/ms
networkThroughput                 100,1ms  thrpt    5    536.800 ±  821.872  ops/ms
networkThroughput              1000,100ms  thrpt    5  30679.754 ± 3737.085  ops/ms
```

master + credit based flow control
```
Benchmark    (number of output channels, flush interval)   Mode  Cnt      Score       Error   Units
networkThroughput                 1,100ms  thrpt    5  49768.778 ± 13329.952  ops/ms
networkThroughput                 100,1ms  thrpt    5  BENCHMARK TIMEOUT! below ~150 ops/ms
networkThroughput              1000,100ms  thrpt    5  27793.594 ±  3428.951  ops/ms
```

credit based + low latency fixes (this PR):
```
Benchmark    (number of output channels, flush interval)   Mode  Cnt      Score       Error   Units
networkThroughput                 1,100ms  thrpt    5  47576.352 ± 12641.958  ops/ms
networkThroughput                 100,1ms  thrpt    5  41898.764 ±  4450.404  ops/ms
networkThroughput              1000,100ms  thrpt    5  27642.259 ±  9086.744  ops/ms
```

## Brief change log

- #4552 and #5314 dependencies
- bunch of hotfixes/prerequisiting fixes
- [FLINK-8582][runtime] Introduce BufferConsumer
- bunch of hotfixes/prerequisiting fixes
- [FLINK-8583] Pass BufferConsumer to subpartitions
- bunch of hotfixes/prerequisiting fixes
- [FLINK-8591][runtime] Pass unfinished bufferConsumers to subpartitions
- some hotfixes

This last one ([FLINK-8591]) is the one commit that actually improves the performance by allowing sender to append a records to a memory segment, while `PartitionRequestQueue` in Netty is busy handling/processing/flushing previous memory segment and when it is blocked for a new credit to arrive.

## Verifying this change

This change is a trivial rework ;)

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): (yes / **no**)
  - The public API, i.e., is any changed class annotated with `@Public(Evolving)`: (yes / **no**)
  - The serializers: (yes / **no** / don't know)
  - The runtime per-record code paths (performance sensitive): (**yes** / no / don't know)
  - Anything that affects deployment or recovery: JobManager (and its components), Checkpointing, Yarn/Mesos, ZooKeeper: (yes / **no** / don't know)
  - The S3 file system connector: (yes / **no** / don't know)

## Documentation

  - Does this pull request introduce a new feature? (yes / **no**)