@StefanRRichter 2018-02-19
Thanks for those very good improvements, I will merge this.